Avoid penalizing decided messages (#413)
This fixes the issues arising with #406. The root cause is as follows:

We send messages anonymously, setting a random author ID through `MessageAuthenticity::RandomAuthor`. Libp2p by default penalizes getting sent messages that originate from ourselves. However, as we do not track under which author we published each message, libp2p simply punishes messages that we have published, identified by message ID. The message ID depends only on the message contents.

The famed "decided message" is generated after a QBFT instance completes - and as it simply is an aggregation of existing signatures and messages, it can be generated by all participants and actually ends up having identical content in many cases. Nodes generate this message at the same time and send it almost simultaneously.

This causes nodes to be penalized for "self" messages, which are not actually self messages.


  Disable penalizing getting sent messages sent by "self" - which may actually be generated by someone else but happens to be identical.