refactor: Revamp startup sequence and check sync status (#372)
We want to be able to listen to the network without an Operator ID, for example to act as an exporter, or to already discover some peers during sync.

Also, we want to make sure to not do anything punishable (invalid network messages, signing slashable objects) if we are out of sync.


  - Add `OwnValidatorId` to allow managers and network to start eagerly
- Provide watcher for synced status, unifying `historic_finished_notify` and `operational_status`
- Avoid sending messages and starting duties for slashable objects if not synced
- Add sync status logging to notifier