feat(database): Add `metadata` table for upgradability and avoiding database switchups (#387)
- closes https://github.com/sigp/anchor/issues/357


  Introduce the `metadata` table where we track the currently used database schema. This will later allow us to implement "migrations", i.e. updating the database schema without requiring the user to resync.

Furthermore, when we create the database, we store the domain type used in order to avoid working with a database that exists, but was previously used on another network.

The existing `block` table is removed and it's column moved to `metadata` in order to have only one table with data that exists only once.