fix(validator_store): Use database directly in validator store (#443)
I noticed a fun bug: When in a single batch (for historic sync) or block (for live sync) a validator is removed and readded again, the validator store does not notice the change. This is bad, as the private key share likely has changed. So we continue signing stuff with the old partial share, while correct nodes sign with the new partial share. This causes signatures to be invalid when recombined.


  First, I was tempted to fix this while maintaining the current structure, but I realized that iterating over the whole database every time and comparing it to everything in the validator store is quite nonsensical if we can use the database directly instead.

So this PR gets rid of the `DashMap`s in the validator store and reads needed info directly from the database state.