fix: Do not punish peers during sync (#513)
During sync, we are happily processing messages. This is fine - it is worthwhile to keep up with the network and start peering ASAP. We will never sign anything during this period, but forward messages.

However, as our state might not be up-to-date yet, validation of valid messages might fail. We do not forward these messages and punish those peers.


  Do not forward invalid messages but do not punish peers as the messages might actually be valid.