fix: preserve order in message containers (#533)
Messages must be stored in the order that they arrive in. This is important for the resulting signatures as even if we have the same messages, if they are not in the expected order this results in different encodings and signatures.

go-ssv just does this through a [slice](https://github.com/ssvlabs/ssv-spec/blob/d4c734909677b5fd3304e1d9220963cc9d794f09/qbft/message_container.go#L9)