fix: properly justify round change (#555)
We were making sure that we had reach consensus on the value that the round changes were preparing, which defeats the purpose of the round change justifications.

Now, we will find the highest prepared and use that data from the round change directly. If for some reason it is prepared and there is no full data, we fall back to our data, otherwise we just use the start value