fix: Validation for prepared round changes before insertion (#528)
If we have prepared a value, we include it in a round change messages and include the quorum of prepare justifications proving that value was prepared in the round_change_justifications field. These round change messages have to have the round_change_justifications prepare messages validated to ensure that these messages actually prove that the value was prepared. Otherwise, you could claim to prepare some value but not properly be able to support/prove it.


  Implement the missing validation and change existing code a bit to be able to reuse `is_valid_prepare_justification_for_round_and_root`.