feat: ssz infrastructure (#580)
This pulls out the ssz infrastructure in #493 into its own PR. Functionally, this should change nothing but it enables hashing for the checks required in the spec tests

This uses new functions `to_variable_list` and `to_variable_list_with_error` which both take the length as a generic to check the upper bound of the data being validated. Allows us to use `From` while also ensuring that we reject invalid length data.

This is the base PR for all of the spec test as we must hash the messages for output comparison against go-ssv


  


Co-Authored-By: Zacholme7 <zacholme@gmail.com>

Co-Authored-By: Zac Holme <79027434+Zacholme7@users.noreply.github.com>