fix: Rpc Error handling (#656)
In my naiviety, I tried specifying just an execution-rpc and got a warning requiring a websockets link.

So I removed the --execution-rpc CLI flag and replaced it with a websockets one. Then anchor wouldn't progress. It simply continued doing a back-off without any real warning about why.

This PR adds a warning log when the RPC fails (which I know it can happen naturally in sync) but it can happen if the RPC is actually dead. This fix, allowed me to realize I needed also an RPC endpoint.

I also fixed a typo.


  


Co-Authored-By: Age Manning <Age@AgeManning.com>